Replace BackendClient with axios-based implementation
so that we can reuse this from the backend.  This builds on the previous commit, making sure that our backend client and REST API tests all work properly.

In the process, we've dramatically simplified the options that get passed to backend clients.
