+++ lib/api/WebApi.js
@@ -5,7 +5,7 @@ import {
   ReportFormat,
   ReportType,
 } from '@/lib/Constants';
-import BackendClient from '@/lib/api/BackendClientNew';
+import AxiosBackendClient from '@/lib/api/AxiosBackendClient';
 import CompositeId from '@/lib/io/CompositeId';
 import AuthState from '@/lib/model/AuthState';
 import Category from '@/lib/model/Category';
@@ -20,8 +20,8 @@ import User from '@/lib/model/User';
 import CentrelineLocation from '@/lib/model/helpers/CentrelineLocation';
 import DateTime from '@/lib/time/DateTime';
 
-const apiClient = new BackendClient('/api');
-const reporterClient = new BackendClient('/reporter');
+const apiClient = new AxiosBackendClient('/api');
+const reporterClient = new AxiosBackendClient('/reporter');
 
 async function deleteStudyRequestComment(csrf, studyRequest, comment) {
   const { id: commentId } = comment;
