+++ lib/test/api/InjectBackendClient.js
@@ -1,36 +1,43 @@
-import BackendClient from '@/lib/api/BackendClient';
+import QueryString from '@/lib/api/QueryString';
 
 class InjectBackendClient {
   constructor(server) {
-    this.backendClient = new BackendClient('');
     this.csrf = null;
     this.server = server;
     this.user = null;
   }
 
-  getInjectOptions(apiUrl, apiOptions) {
+  static getInjectUrl(url, options) {
+    const { data, method = 'GET' } = options;
+    if (data === undefined || method !== 'GET') {
+      return url;
+    }
+    const qs = QueryString.get(data);
+    return `${url}?${qs}`;
+  }
+
+  getInjectOptions(injectUrl, options) {
     const {
-      body,
-      headers,
-      method,
-    } = apiOptions;
+      data,
+      method = 'GET',
+    } = options;
+    const headers = {};
     const injectOptions = {
       headers,
       method,
-      url: apiUrl,
+      url: injectUrl,
     };
-    if (body !== undefined) {
-      injectOptions.payload = body;
-    }
-    if (this.csrf === null) {
-      if (method !== 'GET') {
-        throw new Error('CSRF token required for non-GET API calls');
+    if (method !== 'GET') {
+      if (this.csrf === null) {
+        throw new Error('must provide CSRF token for non-GET requests');
       }
-    } else {
       const { csrf } = this;
       /* eslint-disable-next-line dot-notation */
       headers['Cookie'] = `csrf=${csrf}`;
       headers['X-CSRF-Token'] = csrf;
+      if (data !== undefined) {
+        injectOptions.payload = data;
+      }
     }
     if (this.user !== null) {
       const credentials = this.getAuthCredentials();
@@ -81,10 +88,9 @@ class InjectBackendClient {
     return response;
   }
 
-  async fetch(url, options) {
-    const apiOptions = BackendClient.getFetchOptions(options);
-    const apiUrl = this.backendClient.getFetchUrl(url, apiOptions);
-    const injectOptions = this.getInjectOptions(apiUrl, apiOptions);
+  async fetch(url, options = {}) {
+    const injectUrl = InjectBackendClient.getInjectUrl(url, options);
+    const injectOptions = this.getInjectOptions(injectUrl, options);
     const response = await this.server.inject(injectOptions);
     return this.handleResponse(response);
   }
